autodiff: address PR #600 review — canonical dtype flow + the four-backend assessment
P1 (dtype): forward-mode dispatch returns the rule's primal instead of
re-executing the canonical op, so the derived rules' forced float64
casts silently promoted a float32 trace's result dtype the moment AD was
enabled — and diverged from the displaced tanh/sin/sigmoid hand rules,
which preserved dtype. The derived pair now evaluates value and slope at
the input's dtype: all 13 ops keep float32 primals/tangents/cotangents
for float32 inputs (bit-compatible with the dtype-preserving displaced
rules), and the displaced FACTORY rules' float64 promotion (exp/log/…)
is a deliberate, pinned fix — the old promoting behavior is asserted in
the test so the change stays a decision, not an accident. In-domain
fp64 differentials vs the oracles are unchanged.

P1 (process): the AGENTS.md cross-backend assessment for this shared
numerical-policy change, recorded in all four backend plans under sync
key AD-RETIRE-1-POINTWISE-2026-08-20:
- rocm: parity VALIDATED exact-device on gfx1151 (native JVP compiled +
  spectral backward exec + native-CPU vertical, 24 tests green against
  the retired registry);
- x86: parity VALIDATED on the AVX-512 host (169 compiled-loss backward
  lane tests green); AMX half not applicable (no fleet hardware);
- nvidia: follow-up required on NR2 Pro (expected parity-neutral — both
  sides of every comparison read the same updated reference, and the
  dtype change moves the reference TOWARD the fp32 native lanes — but
  expectation is not evidence);
- apple: follow-up required on the M1 Max (same reasoning, same rule).

+++ python/tessera/autodiff/derivative_contract.py
@@ -152,7 +152,14 @@ def derivative_contract(name: str) -> Optional[DerivativeContract]:
 #
 # What the derived path carries (§8's bar — everything the deleted one did):
 #
-# * **dtype behavior** — float64 casts, identical to the displaced factory.
+# * **dtype behavior** — the CANONICAL dtype flow: value and slope are
+#   evaluated at the input's dtype, so a float32 trace keeps float32
+#   primals and tangents (PR #600 review). This matches the displaced
+#   tanh/sin/sigmoid hand rules bit-for-bit; the displaced FACTORY rules
+#   (exp/log/…) force-promoted to float64, which — because forward-mode
+#   dispatch returns the rule's primal instead of re-executing the
+#   canonical op — silently changed a function's result dtype the moment
+#   AD was enabled. That promotion is deliberately fixed, and pinned.
 # * **evaluation guards** — the displaced rules' domain clamps, carried as
 #   EXPLICIT per-op declarations below instead of buried lambdas. One guard
 #   per op for BOTH modes. This *fixes a measured inconsistency*: the
@@ -180,15 +187,15 @@ def _make_derived_pair(name: str, rec: ScalarRecurrence):
     guard = _DERIVATIVE_EVAL_GUARDS.get(name)
 
     def derived_jvp(primals, tangents, **_):
-        x = np.asarray(primals[0], dtype=np.float64)
-        dx = np.asarray(tangents[0], dtype=np.float64)
+        x = np.asarray(primals[0])
+        dx = np.asarray(tangents[0])
         xg = guard(x) if guard is not None else x
         return np.asarray(value(x)), derivative(xg) * dx
 
     def derived_vjp(dout, x, **_):
-        a = np.asarray(x, dtype=np.float64)
+        a = np.asarray(x)
         ag = guard(a) if guard is not None else a
-        return (derivative(ag) * np.asarray(dout, dtype=np.float64),)
+        return (derivative(ag) * np.asarray(dout),)
 
     derived_jvp._derived_from_datum = name  # type: ignore[attr-defined]
     derived_vjp._derived_from_datum = name  # type: ignore[attr-defined]
